namespace Import -> ChartMogul, backwards compatible
