+++ src/Import/DataSource.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ChartMogul\Import;
+
+/**
+ * @deprecated Use ChartMogul\DataSource
+ */
+class DataSource extends ChartMogul\DataSource {
+
+}
