+++ src/Import/LineItems/AbstractLineItem.php
@@ -4,20 +4,8 @@ namespace ChartMogul\Import\LineItems;
 
 /**
 * @codeCoverageIgnore
-* @property-read string $uuid
+* @deprecated
 */
-abstract class AbstractLineItem extends \ChartMogul\Resource\AbstractModel
+abstract class AbstractLineItem extends \ChartMogul\AbstractLineItem
 {
-
-    protected $uuid;
-
-    public $type;
-    public $amount_in_cents;
-    public $quantity;
-    public $discount_amount_in_cents;
-    public $discount_code;
-    public $tax_amount_in_cents;
-    public $external_id;
-
-    public $invoice_uuid;
 }
