+++ src/Import/LineItems/Subscription.php
@@ -4,17 +4,8 @@ namespace ChartMogul\Import\LineItems;
 
 /**
 * @codeCoverageIgnore
+* @deprecated Use ChartMogul\Subscription
 */
-class Subscription extends AbstractLineItem
-{
-
-    public $type = 'subscription';
-    public $subscription_external_id;
-    public $service_period_start;
-    public $service_period_end;
-    public $cancelled_at;
-    public $prorated;
-
-    protected $subscription_uuid;
-    public $plan_uuid;
+class Subscription extends Subscription{
+    
 }
