+++ src/Import/Plan.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ChartMogul\Import;
+
+/**
+ * @deprecated Use ChartMogul\Plan
+ */
+class Plan extends ChartMogul\Plan {
+
+}
