+++ src/Import/Transactions/AbstractTransaction.php
@@ -2,24 +2,10 @@
 
 namespace ChartMogul\Import\Transactions;
 
-use ChartMogul\Resource\AbstractResource;
-
 /**
  * @property-read string $uuid
+ * @deprecated Use ChartMogul\AbstractTransaction
  */
-abstract class AbstractTransaction extends AbstractResource
-{
-
-    use \ChartMogul\Service\CreateTrait;
-
-    const RESOURCE_PATH = '/v1/import/invoices/:invoice_uuid/transactions';
-
-    protected $uuid;
-
-    public $type;
-    public $date;
-    public $result;
-    public $external_id;
-
-    public $invoice_uuid;
+abstract class AbstractTransaction extends ChartMogul\AbstractTransaction {
+    
 }
