+++ src/Import/Transactions/Payment.php
@@ -3,13 +3,9 @@
 namespace ChartMogul\Import\Transactions;
 
 /**
-* @codeCoverageIgnore
-*/
-class Payment extends AbstractTransaction
-{
-
-    const RESOURCE_NAME= 'Payment Transaction';
-
-
-    public $type = 'payment';
+ * @codeCoverageIgnore
+ * @deprecated Use ChartMogul\Payment
+ */
+class Payment extends ChartMogul\Payment {
+    
 }
