+++ src/LineItems/OneTime.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace ChartMogul\LineItems;
+
+/**
+* @codeCoverageIgnore
+*/
+class OneTime extends AbstractLineItem
+{
+
+    public $type = 'one_time';
+    public $description;
+}
