+++ src/Transactions/Refund.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace ChartMogul\Transactions;
+
+/**
+* @codeCoverageIgnore
+*/
+class Refund extends AbstractTransaction
+{
+
+    const RESOURCE_NAME= 'Refund Transaction';
+
+    public $type = 'refund';
+}
